Login auth step 1 added
The login authentication is added. Actually this update is just the
first step of the OAuth authentication (from asking for a temporary
token to redirecting the user to Fitbit's webpage for authorization).
Please try it out to see how it gonna redirects the user. You may need
to add some style/animation to make the redirection operation better and
more user-friendly.